# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: dashboard
    description: Dashboard summaries and reporting
  - name: leads
    description: Lead management
  - name: companies
    description: Company management
  - name: tasks
    description: Follow-up tasks
  - name: activities
    description: Sales activities
  - name: deals
    description: Deals and revenue
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get dashboard summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /reports:
    get:
      operationId: getReports
      tags: [dashboard]
      summary: Get report snapshots
      responses:
        "200":
          description: Reporting snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportsSnapshot"
  /leads:
    get:
      operationId: getLeads
      tags: [leads]
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: stage
          in: query
          schema:
            type: string
        - name: owner
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "409":
          description: Duplicate lead
  /leads/{id}:
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Lead not found
  /companies:
    get:
      operationId: getCompanies
      tags: [companies]
      responses:
        "200":
          description: Companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a follow-up task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Archive a follow-up task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
  /activities:
    get:
      operationId: getActivities
      tags: [activities]
      responses:
        "200":
          description: Activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
    post:
      operationId: createActivity
      tags: [activities]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityInput"
      responses:
        "201":
          description: Created activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
  /deals:
    get:
      operationId: getDeals
      tags: [deals]
      responses:
        "200":
          description: Deals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deal"
    post:
      operationId: createDeal
      tags: [deals]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealInput"
      responses:
        "201":
          description: Created deal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
  /deals/{id}/payment:
    patch:
      operationId: recordDealPayment
      tags: [deals]
      summary: Record a payment received against a deal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealPayment"
      responses:
        "200":
          description: Updated deal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
        "404":
          description: Deal not found
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Lead:
      type: object
      required: [id, name, company, email, stage, owner, value, source, priority, nextAction, followUpDate, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        company: { type: string }
        email: { type: string }
        phone: { type: string }
        stage: { type: string }
        owner: { type: string }
        value: { type: number }
        source: { type: string }
        priority: { type: string }
        nextAction: { type: string }
        followUpDate: { type: string }
        updatedAt: { type: string }
        isStale: { type: boolean }
    LeadInput:
      type: object
      required: [name, company, email, stage, owner, value, source, priority, nextAction, followUpDate]
      properties:
        name: { type: string, minLength: 1 }
        company: { type: string, minLength: 1 }
        email: { type: string }
        phone: { type: string }
        stage: { type: string }
        owner: { type: string }
        value: { type: number, minimum: 0 }
        source: { type: string }
        priority: { type: string }
        nextAction: { type: string }
        followUpDate: { type: string }
    LeadUpdate:
      type: object
      properties:
        stage: { type: string }
        owner: { type: string }
        value: { type: number, minimum: 0 }
        priority: { type: string }
        nextAction: { type: string }
        followUpDate: { type: string }
    Company:
      type: object
      required: [id, name, industry, leads, openValue, location, health]
      properties:
        id: { type: string }
        name: { type: string }
        industry: { type: string }
        leads: { type: integer }
        openValue: { type: number }
        location: { type: string }
        health: { type: string }
    Task:
      type: object
      required: [id, title, lead, dueDate, assignee, priority, status]
      properties:
        id: { type: string }
        title: { type: string }
        lead: { type: string }
        dueDate: { type: string }
        assignee: { type: string }
        priority: { type: string }
        status: { type: string }
    TaskInput:
      type: object
      required: [title, lead, dueDate, assignee, priority]
      properties:
        title: { type: string, minLength: 1 }
        lead: { type: string, minLength: 1 }
        dueDate: { type: string, minLength: 1 }
        assignee: { type: string, minLength: 1 }
        priority: { type: string }
    TaskUpdate:
      type: object
      properties:
        status: { type: string }
        title: { type: string, minLength: 1 }
        lead: { type: string, minLength: 1 }
        dueDate: { type: string, minLength: 1 }
        assignee: { type: string, minLength: 1 }
        priority: { type: string }
    Activity:
      type: object
      required: [id, type, title, person, detail, timestamp, tone]
      properties:
        id: { type: string }
        type: { type: string }
        title: { type: string }
        person: { type: string }
        detail: { type: string }
        timestamp: { type: string }
        tone: { type: string }
        relatedType: { type: ["string", "null"] }
        relatedTo: { type: ["string", "null"] }
    ActivityInput:
      type: object
      required: [type, title, person, detail]
      properties:
        type: { type: string }
        title: { type: string }
        person: { type: string }
        detail: { type: string }
        relatedType: { type: string }
        relatedTo: { type: string }
    Deal:
      type: object
      required: [id, name, company, value, stage, owner, expectedClose, paid, currency]
      properties:
        id: { type: string }
        name: { type: string }
        company: { type: string }
        value: { type: number }
        stage: { type: string }
        owner: { type: string }
        expectedClose: { type: string }
        paid: { type: number }
        currency: { type: string }
    DealInput:
      type: object
      required: [name, company, value, stage, owner, expectedClose, currency]
      properties:
        name: { type: string }
        company: { type: string }
        value: { type: number, minimum: 0 }
        stage: { type: string }
        owner: { type: string }
        expectedClose: { type: string }
        currency: { type: string }
    DealPayment:
      type: object
      required: [amount]
      properties:
        amount: { type: number, minimum: 0.01 }
    DashboardSummary:
      type: object
      required: [metrics, pipeline, revenue, recentActivities, todayTasks]
      properties:
        metrics:
          type: array
          items:
            $ref: "#/components/schemas/Metric"
        pipeline:
          type: array
          items:
            $ref: "#/components/schemas/PipelineStage"
        revenue:
          type: array
          items:
            $ref: "#/components/schemas/RevenuePoint"
        recentActivities:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
        todayTasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
    ReportsSnapshot:
      type: object
      required: [funnel, sources, owners]
      properties:
        funnel:
          type: array
          items:
            $ref: "#/components/schemas/PipelineStage"
        sources:
          type: array
          items:
            $ref: "#/components/schemas/SourcePerformance"
        owners:
          type: array
          items:
            $ref: "#/components/schemas/OwnerPerformance"
    Metric:
      type: object
      required: [label, value, change, trend, accent]
      properties:
        label: { type: string }
        value: { type: string }
        change: { type: string }
        trend: { type: string }
        accent: { type: string }
    PipelineStage:
      type: object
      required: [label, count, value]
      properties:
        label: { type: string }
        count: { type: integer }
        value: { type: number }
    RevenuePoint:
      type: object
      required: [month, booked, collected]
      properties:
        month: { type: string }
        booked: { type: number }
        collected: { type: number }
    SourcePerformance:
      type: object
      required: [source, leads, won, conversion]
      properties:
        source: { type: string }
        leads: { type: integer }
        won: { type: integer }
        conversion: { type: number }
    OwnerPerformance:
      type: object
      required: [owner, leads, wonValue, activities]
      properties:
        owner: { type: string }
        leads: { type: integer }
        wonValue: { type: number }
        activities: { type: integer }
    DeleteResponse:
      type: object
      required: [success]
      properties:
        success: { type: boolean }