# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Activity {
  id: string;
  type: string;
  title: string;
  person: string;
  detail: string;
  timestamp: string;
  tone: string;
  /** @nullable */
  relatedType?: string | null;
  /** @nullable */
  relatedTo?: string | null;
}