# lib/api-zod/src/generated/types/activityInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityInput {
  type: string;
  title: string;
  person: string;
  detail: string;
  relatedType?: string;
  relatedTo?: string;
}