# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Company {
  id: string;
  name: string;
  industry: string;
  leads: number;
  openValue: number;
  location: string;
  health: string;
}