# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Activity } from './activity';
import type { Metric } from './metric';
import type { PipelineStage } from './pipelineStage';
import type { RevenuePoint } from './revenuePoint';
import type { Task } from './task';

export interface DashboardSummary {
  metrics: Metric[];
  pipeline: PipelineStage[];
  revenue: RevenuePoint[];
  recentActivities: Activity[];
  todayTasks: Task[];
}