# lib/api-zod/src/generated/types/dealInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DealInput {
  name: string;
  company: string;
  /** @minimum 0 */
  value: number;
  stage: string;
  owner: string;
  expectedClose: string;
  currency: string;
}